backend/synthetic_example_generation/requirements.txt
git+https://github.com/IBM/activation-steering.git
accelerate
datasets
langchain
numpy
openai
pandas
python-dotenv
scipy
transformers
torch